Preprocess and Cache Subjects (#14)
- save preprocessed subjects to python pickle files, load into memory at
training time
- improved patch sampling using weighted probability map to avoid
sampling patches near the boundary of the skull that contain mostly
background


* I further sped up training time by about 2x to under 12min for 3k
iterations by preprocessing the data (cropping to the skull, resampling,
adding patch sampler probability mask, rescaling intensity) offline and
saving the resulting “Sample” structs (containing all of the aligned
scans) as python pickle files that I’ll load into memory at training
time.
* I also updated the patch sampling policy to sample patches according
to the probability map shown below. This ensures that patches are
sampled from the interior of the brain so extra padding is not required.
Also, I have defined the unnormalized probability as the log(sum(# tumor
pixels in the 3d patch volume)), so that patches containing more tumor
are sampled more often.
Previously the patch sampling policy would uniformly sample patches
centered on a tumor pixel, which would often cause the sampler to
include a large fraction of background pixels.
* I also tried resampling the volumes by factor ~2x decreasing cropped
volume size from on the order of (140,180,140) to (80, 100, 80) such
that a 64x64x64 patch includes a majority of the volume.